# src/main/resources/api.yml
openapi: 3.0.3
info:
  title: RPG API
  description: API for RPG REST game.
  version: '0.1'
servers:
  - url: http://localhost:8080
paths:
...
  /character:
    post:
      summary: Creates a character.
      operationId: createCharacter
      x-preauth: "hasAuthority('USER')"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CharacterCreateRequest'
      responses:
        '200':
          description: Info about created character
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CharacterLite'
  /character/image/{characterArtwork}:
    get:
      summary: Get full sized character artwork
      operationId: getCharacterArtwork
      parameters:
        - name: characterArtwork
          in: path
          required: true
          description: Artwork enum
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            image/jpeg:
              schema:
                type: string
                format: binary
  /character/thumbnail/{characterArtwork}:
    get:
      summary: Get thumbnail artwork
      operationId: getCharacterArtworkThumbnail
      parameters:
        - name: characterArtwork
          in: path
          required: true
          description: Artwork enum
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            image/jpeg:
              schema:
                type: string
                format: binary
...
  /character/{characterId}:
    get:
      summary: Get user character by id
      operationId: getUserCharacter
      x-preauth: "hasAuthority('USER')"
      parameters:
        - name: characterId
          in: path
          required: true
          description: Character id
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CharacterDetails'
...
  /skill:
    get:
      summary: Get all skills
      operationId: getSkills
      x-preauth: "hasAuthority('ADMIN')"
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SkillLites'
    post:
      summary: Creates new skill.
      operationId: createSkill
      x-preauth: "hasAuthority('ADMIN')"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SkillCreateRequest'
      responses:
        '200':
          description: Info about created skill
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SkillLite'
  /adventure:
    post:
      summary: Creates new adventure.
      operationId: createAdventure
      x-preauth: "hasAuthority('ADMIN')"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdventureCreateRequest'
      responses:
        '200':
          description: Info about created adventure
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdventureLite'
  /adventure/{adventureId}:
    get:
      summary: Get info about adventure by adventure id
      operationId: getAdventure
      x-preauth: "hasAnyAuthority('ADMIN', 'USER')"
      parameters:
        - name: adventureId
          in: path
          required: true
          description: Adventure id
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdventureDetails'
    put:
      summary: Edits adventure.
      operationId: editAdventure
      x-preauth: "hasAuthority('ADMIN')"
      parameters:
        - name: adventureId
          in: path
          required: true
          description: Adventure id
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdventureCreateRequest'
      responses:
        '200':
          description: Info about edited adventure
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdventureLite'
    delete:
      summary: Deletes adventure.
      operationId: deleteAdventure
      x-preauth: "hasAuthority('ADMIN')"
      parameters:
        - name: adventureId
          in: path
          required: true
          description: Adventure id
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Info about deleted adventure
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdventureLite'
  /adventure/{adventureId}/start/{characterId}:
    get:
      summary: Start adventure with character
      operationId: startAdventure
      x-preauth: "hasAuthority('USER')"
      parameters:
        - name: adventureId
          in: path
          required: true
          description: Adventure id
          schema:
            type: integer
            format: int64
        - name: characterId
          in: path
          required: true
          description: Character id that starts adventure
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdventureLite'
  /adventure/{characterId}/end:
    get:
      summary: End adventure with character
      operationId: endAdventure
      x-preauth: "hasAuthority('USER')"
      parameters:
        - name: characterId
          in: path
          required: true
          description: Character id that ends adventure
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdventureLite'
  /adventure/search:
    post:
      summary: Search adventures.
      operationId: findAdventures
      x-preauth: "hasAnyAuthority('ADMIN', 'USER')"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdventureSearchRequest'
      responses:
        '200':
          description: Info about found skills
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdventureBasicPage'
  /skill/search:
    post:
      summary: Search skills.
      operationId: findSkills
      x-preauth: "hasAnyAuthority('ADMIN', 'USER')"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SkillSearchRequest'
      responses:
        '200':
          description: Info about found skills
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SkillBasicPage'
  /skill/{skillId}:
    get:
      summary: Get skill by skill id
      operationId: getSkill
      x-preauth: "hasAnyAuthority('ADMIN', 'USER')"
      parameters:
        - name: skillId
          in: path
          required: true
          description: Skill id
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Info about created skill
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SkillDetails'
  /skill/character/{characterId}:
    get:
      summary: Info about character skills
      operationId: getCharacterSkills
      x-preauth: "hasAuthority('USER')"
      parameters:
        - name: characterId
          in: path
          required: true
          description: Character id that has skills
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Info about learned skill
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CharacterSkillBasics'
  /skill/{skillId}/learn/{characterId}:
    get:
      summary: Learn skill
      operationId: learnSkill
      x-preauth: "hasAuthority('USER')"
      parameters:
        - name: skillId
          in: path
          required: true
          description: Skill id
          schema:
            type: integer
            format: int64
        - name: characterId
          in: path
          required: true
          description: Character id that learns skill
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Info about learned skill
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SkillLite'
...
  /item:
    post:
      summary: Create item.
      operationId: createItem
      x-preauth: "hasAuthority('ADMIN')"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ItemCreateRequest'
      responses:
        '200':
          description: Necessary info about item
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ItemLite'
  /item/search:
    post:
      summary: Search items.
      operationId: findItems
      x-preauth: "hasAnyAuthority('ADMIN', 'USER')"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ItemSearchRequest'
      responses:
        '200':
          description: Info about found items
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ItemLitePage'
  /item/{itemId}/buy/{characterId}:
    get:
      summary: Buy item
      operationId: buyItem
      x-preauth: "hasAuthority('USER')"
      parameters:
        - name: itemId
          in: path
          required: true
          description: Item id
          schema:
            type: integer
            format: int64
        - name: characterId
          in: path
          required: true
          description: Character id
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Info about bought item
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ItemLite'
  /item/potion/buy/{characterId}:
    get:
      summary: Buy potion
      operationId: buyPotion
      x-preauth: "hasAuthority('USER')"
      parameters:
        - name: characterId
          in: path
          required: true
          description: Character id
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: No content
...
  /item/potion/use/{characterId}:
    get:
      summary: Heal character out of combat
      operationId: usePotion
      x-preauth: "hasAuthority('USER')"
      parameters:
        - name: characterId
          in: path
          required: true
          description: Character id
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Statistics info
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatisticsLite'
  /work:
    post:
      summary: Creates new work.
      operationId: createWork
      x-preauth: "hasAuthority('ADMIN')"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WorkCreateRequest'
      responses:
        '200':
          description: Info about created work
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkLite'
  /work/search:
    post:
      summary: Search works.
      operationId: findWorks
      x-preauth: "hasAnyAuthority('ADMIN', 'USER')"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WorkSearchRequest'
      responses:
        '200':
          description: Info about found works
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkLitePage'
  /work/{workId}/start/{characterId}:
    get:
      summary: Start work with character
      operationId: startWork
      x-preauth: "hasAuthority('USER')"
      parameters:
        - name: workId
          in: path
          required: true
          description: Work id
          schema:
            type: integer
            format: int64
        - name: characterId
          in: path
          required: true
          description: Character id that starts adventure
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: No content
  /work/{characterId}/end:
    get:
      summary: End work with character
      operationId: endWork
      x-preauth: "hasAuthority('USER')"
      parameters:
        - name: characterId
          in: path
          required: true
          description: Character id that ends work
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: No content
components:
  schemas:
...
    CharacterCreateRequest:
      type: object
      required:
        - name
        - race
        - sex
        - characterClass
        - artwork
        - statistics
      properties:
        name:
          description: Character name
          type: string
        race:
          description: HUMAN/ELF/DWARF
          type: string
        sex:
          description: MALE/FEMALE
          type: string
        characterClass:
          description: WARRIOR/ROGUE/MAGE
          type: string
        artwork:
          description: Character artwork
          type: string
        statistics:
          description: Character statistics
          $ref: '#/components/schemas/StatisticsUpdateRequest'
...
    SkillCreateRequest:
      type: object
      required:
        - name
        - manaCost
        - type
        - multiplier
        - multiplierPerLevel
        - effectDuration
        - effectDurationPerLevel
        - effectMultiplier
        - effectMultiplierPerLevel
        - goldCost
        - statisticPointsCost
        - characterClass
      properties:
        name:
          description: Skill name
          type: string
        manaCost:
          description: Skill mana cost
          type: integer
        type:
          description: Skill type
          type: string
        multiplier:
          description: Skill multiplier for example totalDamage = normalDamage * multiplier
          type: number
          format: float
        multiplierPerLevel:
          description: Multiplier to add per level
          type: number
          format: float
        effect:
          description: Skill effect
          type: string
        effectDuration:
          description: Skill effect duration in turns
          type: integer
        effectDurationPerLevel:
          description: Skill effect duration to add per level
          type: integer
        effectMultiplier:
          description: Effect multiplier for example Burning -> damagePerTurn = maxEnemyHealth * effectMultiplier
          type: number
          format: float
        effectMultiplierPerLevel:
          description: Effect multiplier to add per level
          type: number
          format: float
        goldCost:
          description: Skill learn gold cost
          type: integer
        statisticPointsCost:
          description: Skill learn statistics points cost
          type: integer
        characterClass:
          description: Character class skill is for
          type: string
    SkillSearchRequest:
      type: object
      properties:
        pagination:
          $ref: '#/components/schemas/Pagination'
        idNotIn:
          description: Skill id not in
          type: array
          items:
            type: integer
            format: int64
        nameLike:
          description: Skill name fragment
          type: string
        skillTypeIn:
          description: Skill type in
          type: array
          items:
            type: string
        skillEffectIn:
          description: Skill effect in
          type: array
          items:
            type: string
        characterClassIn:
          description: Character class skill is for in
          type: array
          items:
            type: string
    AdventureCreateRequest:
      type: object
      required:
        - name
        - adventureLengthInMinutes
        - xpForAdventure
        - goldForAdventure
        - enemy
      properties:
        name:
          description: Adventure name
          type: string
        adventureLengthInMinutes:
          description: Adventure length given in minutes
          type: integer
        xpForAdventure:
          description: XP given to player for adventure
          type: integer
        goldForAdventure:
          description: Gold given to player for adventure
          type: integer
        enemy:
          description: Enemy id that player will fight in an adventure
          type: integer
          format: int64
    AdventureSearchRequest:
      type: object
      properties:
        pagination:
          $ref: '#/components/schemas/Pagination'
        nameLike:
          description: Skill name fragment
          type: string
        adventureTimeGreaterThanOrEqual:
          description: Adventure time in minutes greater than or equal
          type: integer
        adventureTimeLessThanOrEqual:
          description: Adventure time in minutes greater than or equal
          type: integer
        xpGreaterThanOrEqual:
          description: XP for adventure greater than or equal
          type: integer
        xpLessThanOrEqual:
          description: XP for adventure greater than or equal
          type: integer
        goldGreaterThanOrEqual:
          description: Gold for adventure greater than or equal
          type: integer
        goldLessThanOrEqual:
          description: Gold for adventure greater than or equal
          type: integer
        enemyNameLike:
          description: Enemy name fragment
          type: string
...
    ItemCreateRequest:
      type: object
      required:
        - name
        - type
        - power
        - price
      properties:
        name:
          description: Item name
          type: string
        type:
          description: Item type (enum ItemType)
          $ref: '#/components/schemas/ItemType'
        power:
          description: Item attribute power
          type: integer
        price:
          description: Item price
          type: integer
    ItemSearchRequest:
      type: object
      properties:
        pagination:
          $ref: '#/components/schemas/Pagination'
        idNotIn:
          description: Item id not in
          type: array
          items:
            type: integer
            format: int64
        nameLike:
          description: Item name fragment
          type: string
        typeIn:
          description: Item type in (enum ItemType)
          type: array
          items:
            $ref: '#/components/schemas/ItemType'
        powerGreaterThanOrEqual:
          description: Item attribute power greater or equals
          type: integer
        powerLessThanOrEqual:
          description: Item attribute power less or equals
          type: integer
        priceGreaterThanOrEqual:
          description: Item price greater or equals
          type: integer
        priceLessThanOrEqual:
          description: Item price less or equals
          type: integer
    WorkCreateRequest:
      type: object
      required:
        - name
        - wage
        - workMinutes
      properties:
        name:
          description: Work name
          type: string
        wage:
          description: Gold earned for work
          type: integer
        workMinutes:
          description: Work length in minutes
          type: integer
...
    CharacterLite:
      type: object
      required:
        - id
        - name
        - race
        - sex
        - characterClass
        - artwork
      properties:
        id:
          description: Character id
          type: integer
          format: int64
        name:
          description: Character name
          type: string
        race:
          description: Character race
          type: string
        sex:
          description: Character sex
          type: string
        characterClass:
          description: Character class
          type: string
        artwork:
          description: Character artwork
          type: string
...
    CharacterDetails:
      allOf:
        - $ref: '#/components/schemas/CharacterBasic'
        - type: object
          properties:
            skills:
              type: array
              items:
                $ref: '#/components/schemas/SkillDetails'
            equipment:
              $ref: '#/components/schemas/EquipmentDetails'
...
    StatisticsLite:
      type: object
      required:
        - currentXp
        - xpToNextLevel
        - currentLevel
        - maxHp
        - currentHp
        - maxMana
        - currentMana
      properties:
        currentXp:
          description: Character current xp
          type: integer
        xpToNextLevel:
          description: Character xp to next level
          type: integer
        currentLevel:
          description: Character current level
          type: integer
        maxHp:
          description: Character max HP
          type: integer
        currentHp:
          description: Character current HP
          type: integer
        maxMana:
          description: Character max mana
          type: integer
        currentMana:
          description: Character current mana
          type: integer
...
    SkillLite:
      type: object
      required:
        - id
        - name
      properties:
        id:
          description: Skill id
          type: integer
          format: int64
        name:
          description: Skill name
          type: string
    SkillLites:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/SkillLite'
...
    SkillBasicPage:
      allOf:
        - $ref: '#/components/schemas/Page'
        - type: object
          required:
            - content
          properties:
            content:
              type: array
              items:
                $ref: '#/components/schemas/SkillBasic'
    SkillDetails:
      allOf:
        - $ref: '#/components/schemas/SkillBasic'
        - type: object
          required:
            - multiplier
          properties:
            manaCost:
              description: Skill mana cost
              type: integer
            multiplier:
              description: Skill multiplier
              type: number
              format: float
            multiplierPerLevel:
              description: Multiplier to add per level
              type: number
              format: float
            effectDuration:
              description: Skill effect duration in turns
              type: integer
            effectMultiplier:
              description: Effect multiplier
              type: number
              format: float
            effectDurationPerLevel:
              description: Skill effect duration to add per level
              type: integer
            effectMultiplierPerLevel:
              description: Effect multiplier to add per level
              type: number
              format: float
...
    CharacterSkillBasics:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/CharacterSkillBasic'
    AdventureLite:
      type: object
      required:
        - id
        - name
        - adventureTimeInMinutes
      properties:
        id:
          description: Adventure id
          type: integer
          format: int64
        name:
          description: Adventure name
          type: string
        adventureTimeInMinutes:
          description: Adventure time in minutes
          type: integer
...
    AdventureBasicPage:
      allOf:
        - $ref: '#/components/schemas/Page'
        - type: object
          required:
            - content
          properties:
            content:
              type: array
              items:
                $ref: '#/components/schemas/AdventureBasic'
    AdventureDetails:
      allOf:
        - $ref: '#/components/schemas/AdventureBasic'
        - type: object
          properties:
            enemy:
              $ref: '#/components/schemas/EnemyBasic'
...
    ItemLite:
      type: object
      description: Necessary info about item
      required:
        - id
        - name
        - type
        - power
        - price
      properties:
        id:
          description: Item id
          type: integer
          format: int64
        name:
          description: Item name
          type: string
        type:
          description: Item type (enum ItemType)
          type: string
        power:
          description: Item attribute power
          type: integer
        price:
          description: Item price
          type: integer
    ItemLitePage:
      allOf:
        - $ref: '#/components/schemas/Page'
        - type: object
          required:
            - content
          properties:
            content:
              type: array
              items:
                $ref: '#/components/schemas/ItemLite'
...
    WorkLite:
      type: object
      description: Info about potion
      required:
        - id
        - name
        - wage
        - workMinutes
      properties:
        id:
          description: Work id
          type: integer
          format: int64
        name:
          description: Work name
          type: string
        wage:
          description: Gold earned in work
          type: integer
        workMinutes:
          description: Work length in minutes
          type: integer
...